+++ include/metalang99/gen.h
@@ -2,34 +2,8 @@
  * @file
  * Support for C language constructions.
  *
- * # Statement chaining
- *
- * This module exports a bunch of so-called _statement chaining macros_: they expect a statement
- * right after their invocation, and moreover, an invocation of such a macro with a statement
- * afterwards altogether form a single statement.
- *
- * How can this be helpful? Imagine you are writing a macro with the following syntax:
- *
- * @code
- * MY_MACRO(...) { bla bla bla }
- * @endcode
- *
- * Then `MY_MACRO` must expand to a _statement prefix_, i.e. something that expects a statement
- * after itself. One possible solution is to make `MY_MACRO` expand to a sequence of statement
- * chaining macros like this:
- *
- * @code
- * #define MY_MACRO(...) \
- *     ML99_INTRODUCE_VAR_TO_STMT(int x = 5) \
- *         ML99_CHAIN_EXPR_STMT(printf("%d\n", x)) \
- *             and so on...
- * @endcode
- *
- * Here `ML99_CHAIN_EXPR_STMT` accepts the statement formed by `ML99_CHAIN_EXPR_STMT`, which in turn
- * accepts the next statement and so on, until a caller of `MY_MACRO` specifies the final statement,
- * thus completing the chain.
- *
- * @see https://www.chiark.greenend.org.uk/~sgtatham/mp/ for a more involved explanation.
+ * Some decent usage examples can be found in
+ * [datatype99/examples/derive](https://github.com/Hirrolot/datatype99/tree/master/examples/derive).
  */
 
 #ifndef ML99_GEN_H
@@ -42,66 +16,10 @@
 #include <metalang99/list.h>
 #include <metalang99/nat.h>
 #include <metalang99/tuple.h>
-#include <metalang99/util.h>
 #include <metalang99/variadics.h>
 
-#ifdef __COUNTER__
-
-/**
- * Generates a unique identifier @p id in the namespace @p prefix.
- *
- * Let `FOO` be the name of an enclosing macro. Then `FOO_` must be specified for @p prefix, and @p
- * id should be given any meaningful name (this makes debugging easier).
- *
- * # Examples
- *
- * @code
- * #include <metalang99/gen.h>
- *
- * #define FOO(...) FOO_NAMED(ML99_GEN_SYM(FOO_, x), __VA_ARGS__)
- * #define FOO_NAMED(x_sym, ...) \
- *      do { int x_sym = 5; __VA_ARGS__ } while (0)
- *
- * // `x` here will not conflict with the `x` inside `FOO`.
- * FOO({
- *     int x = 7;
- *     printf("x is %d\n", x); // x is 7
- * });
- * @endcode
- *
- * @note Two identical calls to #ML99_GEN_SYM will yield different identifiers, therefore, to refer
- * to the result later, you must save it in an auxiliary macro's parameter, as shown in the example
- * above.
- * @note #ML99_GEN_SYM is defined only if `__COUNTER__` is defined, which must be a macro yielding
- * integral literals starting from 0 incremented by 1 each time it is called. Currently, it is
- * supported at least by Clang, GCC, TCC, and MSVC.
- * @see https://en.wikipedia.org/wiki/Hygienic_macro
- */
-#define ML99_GEN_SYM(prefix, id) ML99_CAT4(prefix, id, _, __COUNTER__)
-
-#endif // __COUNTER__
-
-/**
- * Forces a caller to put a trailing semicolon.
- *
- * It is useful when defining macros, to make them formatted as complete statements.
- *
- * # Examples
- *
- * @code
- * #include <metalang99/gen.h>
- *
- * #define MY_MACRO(fn_name, val_ty, val) \
- *     inline static val_ty fn_name(void) { return val; } \
- *     ML99_TRAILING_SEMICOLON()
- *
- * // Defines a function which always returns 0.
- * MY_MACRO(zero, int, 0);
- * @endcode
- *
- * @note This macro expands to a C declaration, therefore, it can be used outside of functions too.
- */
-#define ML99_TRAILING_SEMICOLON(...) struct ml99_priv_trailing_semicolon
+#include <metalang99/stmt.h> // For backwards compatibility.
+#include <metalang99/util.h> // For backwards compatibility: ML99_GEN_SYM.
 
 /**
  * Puts a semicolon after provided arguments.
@@ -365,120 +283,7 @@
  */
 #define ML99_indexedArgs(n) ML99_call(ML99_indexedArgs, n)
 
-/**
- * A statement chaining macro which introduces several variable definitions to a statement right
- * after its invocation.
- *
- * Variable definitions must be specified as in the first clause of the for-loop.
- *
- * Top-level `break`/`continue` inside a user-provided statement are prohibited.
- *
- * # Example
- *
- * @code
- * #include <metalang99/gen.h>
- *
- * for (int i = 0; i < 10; i++)
- *     ML99_INTRODUCE_VAR_TO_STMT(double x = 5.0, y = 7.0)
- *         if (i % 2 == 0)
- *             printf("i = %d, x = %f, y = %f\n", i, x, y);
- * @endcode
- */
-#define ML99_INTRODUCE_VAR_TO_STMT(...)                                                            \
-    ML99_PRIV_SHADOWS(for (__VA_ARGS__, *ml99_priv_break = (void *)0;                              \
-                           ml99_priv_break != (void *)1;                                           \
-                           ml99_priv_break = (void *)1))
-
-/**
- * The same as #ML99_INTRODUCE_VAR_TO_STMT but deals with a single non-`NULL` pointer.
- *
- * In comparison with #ML99_INTRODUCE_VAR_TO_STMT, this macro generates a little less code. It
- * introduces a pointer to @p ty identified by @p name and initialised to @p init.
- *
- * Top-level `break`/`continue` inside a user-provided statement are prohibited.
- *
- * # Example
- *
- * @code
- * #include <metalang99/gen.h>
- *
- * double x = 5.0, y = 7.0;
- *
- * for (int i = 0; i < 10; i++)
- *     ML99_INTRODUCE_NON_NULL_PTR_TO_STMT(double, x_ptr, &x)
- *         ML99_INTRODUCE_NON_NULL_PTR_TO_STMT(double, y_ptr, &y)
- *             printf("i = %d, x = %f, y = %f\n", i, *x_ptr, *y_ptr);
- * @endcode
- *
- * @note Unlike #ML99_INTRODUCE_VAR_TO_STMT, the generated pointer is guaranteed to be used at least
- * once, meaning that you do not need to suppress the unused variable warning.
- * @note @p init is guaranteed to be executed only once.
- */
-#define ML99_INTRODUCE_NON_NULL_PTR_TO_STMT(ty, name, init)                                        \
-    ML99_PRIV_SHADOWS(for (ty *name = (init); name != (void *)0; name = (void *)0))
-
-/**
- * A statement chaining macro which executes an expression statement derived from @p expr right
- * before the next statement.
- *
- * Top-level `break`/`continue` inside a user-provided statement are prohibited.
- *
- * # Example
- *
- * @code
- * #include <metalang99/gen.h>
- *
- * int x;
- *
- * for(;;)
- *     ML99_CHAIN_EXPR_STMT(x = 5)
- *         ML99_CHAIN_EXPR_STMT(printf("%d\n", x))
- *             puts("abc");
- * @endcode
- */
-#define ML99_CHAIN_EXPR_STMT(expr)                                                                 \
-    ML99_PRIV_SHADOWS(for (int ml99_priv_expr_stmt_break = ((expr), 0);                            \
-                           ml99_priv_expr_stmt_break != 1;                                         \
-                           ml99_priv_expr_stmt_break = 1))
-
-/**
- * The same as #ML99_CHAIN_EXPR_STMT but executes @p expr **after** the next statement.
- */
-#define ML99_CHAIN_EXPR_STMT_AFTER(expr)                                                           \
-    ML99_PRIV_SHADOWS(for (int ml99_priv_expr_stmt_after_break = 0;                                \
-                           ml99_priv_expr_stmt_after_break != 1;                                   \
-                           ((expr), ml99_priv_expr_stmt_after_break = 1)))
-
-/**
- * A statement chaining macro which suppresses the "unused X" warning right before a statement after
- * its invocation.
- *
- * Top-level `break`/`continue` inside a user-provided statement are prohibited.
- *
- * # Example
- *
- * @code
- * #include <metalang99/gen.h>
- *
- * int x, y;
- *
- * for(;;)
- *     ML99_SUPPRESS_UNUSED_BEFORE_STMT(x)
- *         ML99_SUPPRESS_UNUSED_BEFORE_STMT(y)
- *             puts("abc");
- * @endcode
- *
- * @deprecated Use `ML99_CHAIN_EXPR_STMT((void)expr)` instead.
- */
-#define ML99_SUPPRESS_UNUSED_BEFORE_STMT(expr) ML99_CHAIN_EXPR_STMT((void)expr)
-
 #ifndef DOXYGEN_IGNORE
-
-#define ML99_PRIV_SHADOWS(...)                                                                     \
-    ML99_CLANG_PRAGMA("clang diagnostic push")                                                     \
-    ML99_CLANG_PRAGMA("clang diagnostic ignored \"-Wshadow\"")                                     \
-    __VA_ARGS__                                                                                    \
-    ML99_CLANG_PRAGMA("clang diagnostic pop")
 
 #define ML99_semicoloned_IMPL(...)                    v(__VA_ARGS__;)
 #define ML99_braced_IMPL(...)                         v({__VA_ARGS__})