+++ scripts/test.sh
@@ -8,4 +8,7 @@ cmake --build .
 if [[ "$OSTYPE" == "linux-gnu" ]]; then
     echo "Testing ./gen ..."
     ./gen
+
+    echo "Testing ./stmt ..."
+    ./stmt
 fi