+++ tests/gen.c
@@ -1,8 +1,8 @@
 #include <metalang99/assert.h>
 #include <metalang99/gen.h>
+#include <metalang99/ident.h>
 #include <metalang99/list.h>
 #include <metalang99/tuple.h>
-#include <metalang99/util.h>
 
 #include <assert.h>
 
@@ -19,148 +19,6 @@ static void
 }
 
 int main(void) {
-
-    // ML99_GEN_SYM
-    {
-
-#define TEST(...) TEST_NAMED(ML99_GEN_SYM(TEST_, x), __VA_ARGS__)
-#define TEST_NAMED(x_sym, ...)                                                                     \
-    do {                                                                                           \
-        int x_sym = 5;                                                                             \
-        (void)x_sym;                                                                               \
-        __VA_ARGS__                                                                                \
-    } while (0)
-
-        // `x` here will not conflict the the `x` inside `TEST`.
-        TEST(int x = 123; (void)x;);
-
-#undef TEST
-#undef TEST_NAMED
-
-// Two identical calls to `ML99_GEN_SYM` must yield different identifiers.
-#define TEST(x1_sym, x2_sym)                                                                       \
-    do {                                                                                           \
-        int x1_sym, x2_sym;                                                                        \
-        (void)x1_sym;                                                                              \
-        (void)x2_sym;                                                                              \
-    } while (0)
-
-        TEST(ML99_GEN_SYM(TEST_, x), ML99_GEN_SYM(TEST_, x));
-
-#undef TEST
-    }
-
-    // ML99_TRAILING_SEMICOLON
-    {
-        ML99_TRAILING_SEMICOLON();
-        ML99_TRAILING_SEMICOLON(~, ~, ~);
-    }
-
-    // ML99_INTRODUCE_VAR_TO_STMT
-    {
-        if (1)
-            ML99_INTRODUCE_VAR_TO_STMT(int x = 5, y = 7) {
-                assert(5 == x);
-                assert(7 == y);
-            }
-    }
-
-    // ML99_INTRODUCE_NON_NULL_PTR_TO_STMT
-    {
-        int x = 5, y = 7;
-
-        // clang-format off
-        if (1)
-            ML99_INTRODUCE_NON_NULL_PTR_TO_STMT(int, x_ptr, &x)
-                ML99_INTRODUCE_NON_NULL_PTR_TO_STMT(int, y_ptr, &y) {
-                    assert(x == *x_ptr);
-                    assert(y == *y_ptr);
-                }
-        // clang-format on
-    }
-
-    // ML99_CHAIN_EXPR_STMT
-    {
-        int x, y;
-
-        // clang-format off
-        if (1)
-            ML99_CHAIN_EXPR_STMT(x = 1)
-                ML99_CHAIN_EXPR_STMT(y = 2) {
-                    assert(1 == x);
-                    assert(2 == y);
-                }
-        // clang-format on
-
-        // Test -Wunused suppression via ML99_CHAIN_EXPR_STMT.
-        int z;
-
-        if (1)
-            ML99_CHAIN_EXPR_STMT((void)z);
-    }
-
-    // ML99_CHAIN_EXPR_STMT_AFTER
-    {
-        int x = 5, y = 7;
-
-        if (1) {
-            assert(5 == x);
-            assert(7 == y);
-
-            ML99_CHAIN_EXPR_STMT_AFTER(x = 1) {
-                assert(5 == x);
-                assert(7 == y);
-
-                ML99_CHAIN_EXPR_STMT_AFTER(y = 2) {
-                    assert(5 == x);
-                    assert(7 == y);
-                }
-
-                assert(5 == x);
-                assert(2 == y);
-            }
-
-            assert(1 == x);
-            assert(2 == y);
-        }
-    }
-
-    // ML99_SUPPRESS_UNUSED_BEFORE_STMT
-    {
-        int x, y;
-
-        // clang-format off
-        if (1)
-            ML99_SUPPRESS_UNUSED_BEFORE_STMT(x)
-                ML99_SUPPRESS_UNUSED_BEFORE_STMT(y)
-                    ;
-        // clang-format on
-    }
-
-    // ClangFormat breaks with the following sequence of statements so they are put into a macro.
-    // clang-format off
-
-#define STMT_CHAINING \
-    ML99_INTRODUCE_VAR_TO_STMT(int x = 5) \
-        ML99_INTRODUCE_NON_NULL_PTR_TO_STMT(int, x_ptr, &x) { \
-            assert(x == *x_ptr); \
- \
-            ML99_CHAIN_EXPR_STMT(x = 7) \
-                ML99_INTRODUCE_VAR_TO_STMT(int y = 5) \
-                    ML99_SUPPRESS_UNUSED_BEFORE_STMT(y) { \
-                        ML99_CHAIN_EXPR_STMT_AFTER(x = 123) { \
-                            assert(7 == x); \
-                        } \
- \
-                        assert(123 == x); \
-                    } \
-        }
-
-    // clang-format on
-
-    { STMT_CHAINING }
-
-#undef STMT_CHAINING
 
     // ML99_braced
     { struct TestBraced ML99_EVAL(ML99_braced(v(int a, b, c;))); }
@@ -209,9 +67,7 @@ int main(void) {
     // ML99_prefixedBlock
     {
 
-        ML99_EVAL(ML99_prefixedBlock(
-            v(ML99_INTRODUCE_VAR_TO_STMT(int x = 5)),
-            v(assert(x == 5); goto end_prefixed_block;)))
+        ML99_EVAL(ML99_prefixedBlock(v(if (1)), v(goto end_prefixed_block;)))
 
         // Unreachable code.
         assert(0);